+++ dico_toolbox/volume.py
@@ -1,0 +1,21 @@
+from soma import aims as _aims
+
+def new_aims_volume_like(vol):
+    """Create a new empty aims.Volume with the same shape as vol
+
+    :param vol: the volume data
+    :type vol: np.ndarray
+    :return: an empty Aims Volume object of the same shape as vol
+    :rtype: aims.Volume
+    """
+    # set same dimensions and data type
+    new_vol = _aims.Volume(
+        vol.getSizeX(),
+        vol.getSizeY(),
+        vol.getSizeZ(),
+        vol.getSizeT(),
+        dtype=vol.arraydata().dtype
+    )
+    # copy the header
+    new_vol.header().update(vol.header())
+    return new_vol
